[#102] Add ComposeCryptoApp
