+++ app/src/main/java/co/nimblehq/compose/crypto/ui/screens/MainActivity.kt
@@ -6,7 +6,7 @@ import androidx.activity.compose.setContent
 import androidx.core.view.WindowCompat
 import co.nimblehq.compose.crypto.domain.usecase.IsNetworkConnectedUseCase
 import co.nimblehq.compose.crypto.rememberCryptoAppState
-import co.nimblehq.compose.crypto.ui.navigation.AppNavigation
+import co.nimblehq.compose.crypto.ui.navigation.ComposeCryptoApp
 import co.nimblehq.compose.crypto.ui.theme.ComposeTheme
 import co.nimblehq.compose.crypto.util.DispatchersProvider
 import dagger.hilt.android.AndroidEntryPoint
@@ -27,7 +27,7 @@ class MainActivity : ComponentActivity() {
         WindowCompat.setDecorFitsSystemWindows(window, false)
         setContent {
             ComposeTheme {
-                AppNavigation(
+                ComposeCryptoApp(
                     cryptoAppState = rememberCryptoAppState(
                         isNetworkConnectedUseCase = isNetworkConnectedUseCase,
                         dispatchersProvider = dispatchersProvider,
